+++ parapet-cats-effect/src/main/scala/io/parapet/cats/CatsEffectParapetRuntime.scala
@@ -4,8 +4,7 @@ import cats.effect.{FiberIO, IO, Outcome}
 import cats.effect.unsafe.implicits.global
 import cats.syntax.all.*
 import io.parapet.ParApp
-import io.parapet.core.Parallel
-import io.parapet.effect.{Effect, EffectFiber}
+import io.parapet.effect.{Effect, EffectFiber, Parallel}
 import io.parapet.runtime.SchedulerRuntime
 
 import java.util.concurrent.atomic.AtomicInteger
