+++ parapet-core/src/main/scala/io/parapet/core/DslInterpreter.scala
@@ -4,7 +4,7 @@ import io.parapet.runtime.Context.ProcessState
 import io.parapet.core.Dsl.*
 import io.parapet.runtime.Scheduler.{Deliver, ProcessQueueIsFull}
 import io.parapet.core.exceptions.RecoveryContractViolation
-import io.parapet.effect.{Deferred, Effect}
+import io.parapet.effect.{Deferred, Effect, Fiber}
 import io.parapet.effect.Monad.*
 import io.parapet.free.{FunctionK, ~>}
 import io.parapet.runtime.{Context, Envelope, EventTransformer, Scheduler, Scope}
