+++ parapet-core/src/main/scala/io/parapet/effect/Clock.scala
@@ -1,4 +1,6 @@
-package io.parapet.core
+package io.parapet.effect
+
+import io.parapet.effect.Clock
 
 import scala.concurrent.duration.FiniteDuration
