+++ parapet-core/src/test/scala/io/parapet/core/LockSpec.scala
@@ -1,6 +1,7 @@
 package io.parapet.core
 
 import io.parapet.core.TestUtils.TestIO
+import io.parapet.effect.Lock
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers.*
